Migrated over the INFO parsing to json.
I also added the logic to copy files into place. This all seems to be working, I just need to add more content to the infos' conditions, and add support for linking to other topics.

The filename handling of the input .fuz files could be improved, since often the text used in VASynth isn't quite the readable text you want for the info response.